Fix compatibility with Pillow 13 (#136)
Pillow allows you to save 32-bit tensors as PNG. However, PNG doesn't
support 32-bit images, so doing this actually converts the data to 16-bit.
Apparently Pillow 13 is going to stop supporting this implicit conversion,
so we have to do it ourselves.

This is technically a change in behavior, because Pillow handles
out-of-range values by clipping them to 0xffff, while Datumaro's `save_image`
just chops off the high bits (via `ndarray.astype`). I don't think it really
matters, though, since data is lost either way.

+++ src/datumaro/plugins/data_formats/cityscapes.py
@@ -1,5 +1,5 @@
 # Copyright (C) 2020-2023 Intel Corporation
-# Copyright (C) 2022 CVAT.ai Corporation
+# Copyright (C) 2022-2026 CVAT.ai Corporation
 #
 # SPDX-License-Identifier: MIT
 
@@ -522,7 +522,7 @@ class CityscapesExporter(Exporter):
                     inst_mask_path,
                     compiled_mask.instance_mask,
                     apply_colormap=False,
-                    dtype=np.int32,
+                    dtype=np.uint16,
                 )
         self.save_label_map()
